Separate login form into its own LoginCard
ask_user calls that carry `fields` (credential forms) now render via a
dedicated LoginCard: a pop-up to fill in, and a compact status line in the
transcript — never the credentials or the raw answer. Plain ask_user
(no fields) and the QR flow are unchanged. This removes the password leak
from the ask_user card's RESPONSE block (it rendered the raw answer).
