+++ components/chat/messages/tools/index.ts
@@ -4,6 +4,7 @@ export { FileDiffCard } from './file-diff-card'
 export { GrepCard } from './grep-card'
 export { GenericCard } from './generic-card'
 export { AskUserCard } from './ask-user-card'
+export { LoginCard } from './login-card'
 export { BackgroundCard } from './background-card'
 export { PlanCard } from './plan-card'
 export { GuideCard } from './guide-card'
